+++ R/namespace.R
@@ -1,0 +1,4 @@
+#' @import httr2
+#' @import jsonlite
+#' @import utils
+#' @import base64enc